Code Cleanup - Wrapped PageShell for Foreground pages
