+++ src/v2/containers/Foreground/PageShell/PageShell.module.scss
@@ -1,0 +1,65 @@
+@import '../../../common/global.module.scss';
+
+.PageShell {
+    @include fullScreen;
+    @include posFixed($z: 9999);
+    color: lightgrey;
+    transition: transform 0.4s ease;
+    display: grid;
+    grid-template-columns: 1fr 2fr 1fr;
+    overflow: hidden;
+}
+
+.page {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+    grid-column: 2 / 3;
+
+    header {
+        height: max(140px, 18vh);
+    }
+
+    section {
+        height: max(620px, 82vh);
+        overflow-y: scroll;
+        -ms-overflow-style: none;  /* Internet Explorer 10+ */
+        scrollbar-width: none;  /* Firefox */
+    }
+
+    section::-webkit-scrollbar { 
+        display: none;  /* Safari and Chrome */
+    }
+}
+
+.sides {
+    display: flex;
+    height: max(140px, 18vh);
+}
+
+.uiBtns {
+    background: transparent;
+    border: none;
+    user-select: none;
+    outline: none;
+    opacity: 0.8;
+    transition: 0.2s ease;
+
+    &:hover {
+        opacity: 1;
+    }
+}
+
+.main {
+    display: grid;
+    grid-template-rows: repeat(1, 100vh);
+    overflow-y: visible;
+    transition: transform 0.4s ease;
+    grid-column: 2 / 3;
+}
+
+.RootSettings {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    row-gap: 24px;
+}
